Replaced the FFTs in h16k and h1k with the most recent ska-sa repo.
The ska-sa repo commit is 3e5787f and has a number of fixes that
we are now implemented in our designs.  Previously, there was add
latency greater than 1 issue which broke the FFT and an 1/8 spectra
glitch.

FPGA designs h1k, h16k, and l1 have been recompiled and quickly
tested.  The bof files have been added to the bofs directory and
old boffiles have been moved to a bofs/old folder.

l8 compile and test is underway, stay tuned...